backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pillow
numpy
